refactor results to drl result
