--- src/core/interfaces/factories/repositories/ResultRepositoryFactoryInterface.php
@@ -1,12 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace DrlArchive\core\interfaces\factories\repositories;
-
-
-use DrlArchive\core\interfaces\repositories\ResultRepositoryInterface;
-
-interface ResultRepositoryFactoryInterface
-{
-    public function create(): ResultRepositoryInterface;
-}
