+++ src/core/interfaces/repositories/DrlResultRepositoryInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace DrlArchive\core\interfaces\repositories;
 
 
-interface ResultRepositoryInterface
+interface DrlResultRepositoryInterface
 {
     const UNABLE_TO_INSERT_EXCEPTION = 2301;
 }
