+++ test/core/entities/DrlResultEntityTest.php
@@ -4,25 +4,31 @@ declare(strict_types=1);
 namespace test\core\entities;
 
 
-use core\entities\ResultEntity;
+use DrlArchive\core\entities\AbstractResultEntity;
+use DrlArchive\core\entities\DrlResultEntity;
 use DrlArchive\core\entities\Entity;
 use DrlArchive\core\entities\EventEntity;
 use DrlArchive\core\entities\TeamEntity;
 use PHPUnit\Framework\TestCase;
 
-class ResultEntityTest extends TestCase
+class DrlResultEntityTest extends TestCase
 {
     public function testInstantiation(): void
     {
+        $entity = new DrlResultEntity();
         $this->assertInstanceOf(
             Entity::class,
-            new ResultEntity()
+            $entity
+        );
+        $this->assertInstanceOf(
+            AbstractResultEntity::class,
+            $entity
         );
     }
 
     public function testIdProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setId(4);
 
         $this->assertEquals(
@@ -33,7 +39,7 @@ class ResultEntityTest extends TestCase
 
     public function testPositionProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setPosition(3);
         $this->assertEquals(
             3,
@@ -43,7 +49,7 @@ class ResultEntityTest extends TestCase
 
     public function testPealNumberProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setPealNumber(3);
         $this->assertEquals(
             3,
@@ -53,7 +59,7 @@ class ResultEntityTest extends TestCase
 
     public function testFaultsProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setFaults(20.5);
         $this->assertEquals(
             20.5,
@@ -63,7 +69,7 @@ class ResultEntityTest extends TestCase
 
     public function testPointsProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setPoints(20);
         $this->assertEquals(
             20,
@@ -73,7 +79,7 @@ class ResultEntityTest extends TestCase
 
     public function testTeamProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setTeam(new TeamEntity());
         $this->assertInstanceOf(
             TeamEntity::class,
@@ -83,7 +89,7 @@ class ResultEntityTest extends TestCase
 
     public function testEventProperty(): void
     {
-        $result = new ResultEntity();
+        $result = new DrlResultEntity();
         $result->setEvent(new EventEntity());
         $this->assertInstanceOf(
             EventEntity::class,
